Extract common functionality
